# platform/openapi.yaml
openapi: 3.0.3
info:
  title: Platform Service API
  description: |
    User authentication, organization management, invitation system,
    and proxy endpoints to pipeline/plugin microservices.
    Supports JWT and OAuth (Google) authentication.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://platform:3000
    description: Internal Docker network
  - url: http://localhost:3000
    description: Local development

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /user/profile:
    get:
      summary: Get current user profile
      operationId: getProfile
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/UserProfile'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      summary: Update current user profile
      operationId: updateProfile
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                email: { type: string, format: email }
              minProperties: 1
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/UserProfile'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /user/account:
    delete:
      summary: Delete current user account
      operationId: deleteAccount
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /user/change-password:
    post:
      summary: Change password
      operationId: changePassword
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password }
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /user/generate-token:
    post:
      summary: Generate new API token pair
      operationId: generateToken
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tokens generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      accessToken: { type: string }
                      refreshToken: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------------------
  # Organization
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT refresh token from login response
...
  schemas:
    SuccessResponse:
      type: object
      required: [success, statusCode]
      properties:
        success: { type: boolean, enum: [true] }
        statusCode: { type: integer }
    MessageResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success: { type: boolean, enum: [true] }
        statusCode: { type: integer }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, statusCode, message]
      properties:
        success: { type: boolean, enum: [false] }
        statusCode: { type: integer }
        message: { type: string }
        code: { type: string }
...
    UserProfile:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [user, admin] }
        isEmailVerified: { type: boolean }
        organizationId: { type: string, nullable: true }
        organizationName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Resource conflict (duplicate)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - bearerAuth: []
...